+++ tests/RePraxis.Tests/TestRePraxis.cs
@@ -1,7 +1,27 @@
 namespace RePraxis.Tests;
 
+[TestFixture]
 public class Tests
 {
+	private RePraxisDatabase _db;
+
+	[SetUp]
+	public void SetUp()
+	{
+		_db = new RePraxisDatabase();
+
+		_db.Insert( "astrid.relationships.jordan.reputation!30" );
+		_db.Insert( "astrid.relationships.jordan.tags.rivalry" );
+		_db.Insert( "astrid.relationships.britt.reputation!-10" );
+		_db.Insert( "astrid.relationships.britt.tags.ex_lover" );
+		_db.Insert( "astrid.relationships.lee.reputation!20" );
+		_db.Insert( "astrid.relationships.lee.tags.friend" );
+		_db.Insert( "britt.relationships.player.tags.spouse" );
+		_db.Insert( "player.relationships.jordan.reputation!-20" );
+		_db.Insert( "player.relationships.jordan.tags.enemy" );
+		_db.Insert( "player.relationships.britt.tags.spouse" );
+	}
+
 	/// <summary>
 	/// Test that create, retrieval, updating, and deletion (CRUD)
 	/// </summary>
@@ -54,155 +74,238 @@ public class Tests
 	}
 
 	[Test]
-	public void TestQuery()
+	public void TestAssertExpressionNoVars()
 	{
-		var db = new RePraxisDatabase();
+		var query = new DBQuery()
+					.Where( "astrid.relationships.britt" );
 
-		db.Insert( "astrid.relationships.jordan.reputation!30" );
-		db.Insert( "astrid.relationships.jordan.tags.rivalry" );
-		db.Insert( "astrid.relationships.britt.reputation!-10" );
-		db.Insert( "astrid.relationships.britt.tags.ex_lover" );
-		db.Insert( "astrid.relationships.lee.reputation!20" );
-		db.Insert( "astrid.relationships.lee.tags.friend" );
-		db.Insert( "britt.relationships.player.tags.spouse" );
-		db.Insert( "player.relationships.jordan.reputation!-20" );
-		db.Insert( "player.relationships.jordan.tags.enemy" );
-		db.Insert( "player.relationships.britt.tags.spouse" );
+		var result = query.Run( _db );
 
-		// Relational expression with a single variable
-		var r0 = new DBQuery()
-			.Where( "astrid.relationships.?other.reputation!?r" )
-			.Where( "gte ?r 10" )
-			.Run( db );
+		Assert.That( result.Success, Is.EqualTo( true ) );
+		Assert.That( result.Bindings.Length, Is.EqualTo( 0 ) );
+	}
 
-		Assert.That( r0.Success, Is.EqualTo( true ) );
-		Assert.That( r0.Bindings.Length, Is.EqualTo( 2 ) );
-
-		// Relational expression with a single variable
-		var r0_b = new DBQuery()
-			.Where( "astrid.relationships.?other.reputation!?r" )
-			.Where( "gte ?r 10" )
-			.Run( db, new Dictionary<string, string>() { { "?other", "lee" } } );
-
-		Assert.That( r0_b.Success, Is.EqualTo( true ) );
-		Assert.That( r0_b.Bindings.Length, Is.EqualTo( 1 ) );
-
-		// Relational expression with multiple variables
-		var r1 = new DBQuery()
-			.Where( "?A.relationships.?other.reputation!?r" )
-			.Where( "lte ?r 0" )
-			.Run( db );
-
-		Assert.That( r1.Success, Is.EqualTo( true ) );
-		Assert.That( r1.Bindings.Length, Is.EqualTo( 2 ) );
-
-		// Assertion expression without variables
-		var r2 = new DBQuery()
-			.Where( "astrid.relationships.britt" )
-			.Run( db );
-
-		Assert.That( r2.Success, Is.EqualTo( true ) );
-		Assert.That( r2.Bindings.Length, Is.EqualTo( 0 ) );
-
+	[Test]
+	public void TestFailingAssertExpressionNoVars()
+	{
 		// Failing assertion without variables
-		var r3 = new DBQuery()
-			.Where( "astrid.relationships.haley" )
-			.Run( db );
+		var query = new DBQuery()
+			.Where( "astrid.relationships.haley" );
 
-		Assert.That( r3.Success, Is.EqualTo( false ) );
-		Assert.That( r3.Bindings.Length, Is.EqualTo( 0 ) );
+		var result = query.Run( _db );
 
-		// Compound query with multiple variables
-		var r4 = new DBQuery()
-			.Where( "?speaker.relationships.?other.reputation!?r0" )
-			.Where( "gt ?r0 10" )
-			.Where( "player.relationships.?other.reputation!?r1" )
-			.Where( "lt ?r1 0" )
-			.Where( "neq ?speaker player" )
-			.Run( db );
+		Assert.That( result.Success, Is.EqualTo( false ) );
+		Assert.That( result.Bindings.Length, Is.EqualTo( 0 ) );
+	}
 
-		Assert.That( r4.Success, Is.EqualTo( true ) );
-		Assert.That( r4.Bindings.Length, Is.EqualTo( 1 ) );
+	[Test]
+	public void TestGteExpression()
+	{
+		var query = new DBQuery()
+			.Where( "astrid.relationships.?other.reputation!?r" )
+			.Where( "gte ?r 10" );
 
+		var result = query.Run( _db );
+
+		Assert.That( result.Success, Is.EqualTo( true ) );
+		Assert.That( result.Bindings.Length, Is.EqualTo( 2 ) );
+	}
+
+	[Test]
+	public void TestGteExpressionWithBindings()
+	{
+		var query = new DBQuery()
+			.Where( "astrid.relationships.?other.reputation!?r" )
+			.Where( "gte ?r 10" );
+
+		var result = query.Run( _db, new Dictionary<string, object>() { { "?other", "lee" } } );
+
+		Assert.That( result.Success, Is.EqualTo( true ) );
+		Assert.That( result.Bindings.Length, Is.EqualTo( 1 ) );
+	}
+
+	[Test]
+	public void TestLteWithMultipleVars()
+	{
+		// Relational expression with multiple variables
+		var query = new DBQuery()
+			.Where( "?A.relationships.?other.reputation!?r" )
+			.Where( "lte ?r 0" );
+
+		var result = query.Run( _db );
+
+		Assert.That( result.Success, Is.EqualTo( true ) );
+		Assert.That( result.Bindings.Length, Is.EqualTo( 2 ) );
+	}
+
+	[Test]
+	public void TestNotExpression()
+	{
 		// Check that a sentence without variables is not true within the database
-		var r5 = new DBQuery()
-			.Where( "not player.relationships.jordan.reputation!30" )
-			.Run( db );
+		var query = new DBQuery()
+			.Where( "not player.relationships.jordan.reputation!30" );
 
-		Assert.That( r5.Success, Is.EqualTo( true ) );
+		var result = query.Run( _db );
 
+		Assert.That( result.Success, Is.EqualTo( true ) );
+	}
+
+	[Test]
+	public void TestNotExpressionWithVars()
+	{
+		// For all relationships astrid has with all ?others,
+		// no relationship has a reputation of 15
+		var query = new DBQuery()
+			.Where( "not astrid.relationships.?other.reputation!15" );
+
+		var result = query.Run( _db );
+
+		Assert.That( result.Success, Is.EqualTo( true ) );
+	}
+
+	[Test]
+	public void TestNeqExpressionWithVars()
+	{
+		// For all relationships astrid has with an ?other,
+		// filter for those where reputation is not 30
+		var query = new DBQuery()
+			.Where( "astrid.relationships.?other.reputation!?rep" )
+			.Where( "neq ?rep 30" );
+
+		var result = query.Run( _db );
+
+		Assert.That( result.Success, Is.EqualTo( true ) );
+		Assert.That( result.Bindings.Length, Is.EqualTo( 2 ) ); // britt and lee
+	}
+
+	[Test]
+	public void TestNotExpressionWithBindings()
+	{
 		// Given that ?other is jordan, is the statement still not true?
-		var r6 = new DBQuery()
-			.Where( "not player.relationships.?other.reputation!30" )
-			.Run( db, new Dictionary<string, string>()
+		var query = new DBQuery()
+			.Where( "not player.relationships.?other.reputation!30" );
+
+		var result = query.Run( _db, new Dictionary<string, object>()
 			{
 				{"?other", "jordan"}
 			} );
 
-		Assert.That( r6.Success, Is.EqualTo( true ) );
+		Assert.That( result.Success, Is.EqualTo( true ) );
+	}
 
-		// For all relationships astrid has with all ?others,
-		// no relationship has a reputation of 15
-		var r7 = new DBQuery()
-			.Where( "not astrid.relationships.?other.reputation!15" )
-			.Run( db );
-
-		Assert.That( r7.Success, Is.EqualTo( true ) );
-
-		// For all relationships astrid has with an ?other,
-		// filter for those where reputation is not 30
-		var r7_b = new DBQuery()
-			.Where( "astrid.relationships.?other.reputation!?rep" )
-			.Where( "neq ?rep 30" )
-			.Run( db );
-
-		Assert.That( r7_b.Success, Is.EqualTo( true ) );
-		Assert.That( r7_b.Bindings.Length, Is.EqualTo( 2 ) ); // britt and lee
-
+	[Test]
+	public void TestCompoundNotQueries()
+	{
+		DBQuery query;
+		QueryResult result;
 
 		// For all relationships astrid has with an ?other
 		// filter for those where reputation is not 30
-		var r8 = new DBQuery()
+		query = new DBQuery()
 			.Where( "astrid.relationships.?other" )
-			.Where( "not astrid.relationships.?other.reputation!30" )
-			.Run( db );
+			.Where( "not astrid.relationships.?other.reputation!30" );
 
-		Assert.That( r8.Success, Is.EqualTo( true ) );
-		Assert.That( r8.Bindings.Length, Is.EqualTo( 2 ) ); // britt and lee
+		result = query.Run( _db );
+
+		Assert.That( result.Success, Is.EqualTo( true ) );
+		Assert.That( result.Bindings.Length, Is.EqualTo( 2 ) ); // britt and lee
 
 		// For all relationships astrid has with an ?other
 		// filter for those where reputation from astrid to ?other is not 30
 		// and ?other does not have a relationship with a spouse tag
-		var r9 = new DBQuery()
+		query = new DBQuery()
 			.Where( "astrid.relationships.?other" )
 			.Where( "not astrid.relationships.?other.reputation!30" )
-			.Where( "not ?other.relationships.?others_spouse.tags.spouse" )
-			.Run( db );
+			.Where( "not ?other.relationships.?others_spouse.tags.spouse" );
 
-		Assert.That( r9.Success, Is.EqualTo( true ) );
-		Assert.That( r9.Bindings.Length, Is.EqualTo( 1 ) ); // lee
+		result = query.Run( _db );
+
+		Assert.That( result.Success, Is.EqualTo( true ) );
+		Assert.That( result.Bindings.Length, Is.EqualTo( 1 ) ); // lee
 
 		// For all relationships astrid has with an ?other
 		// filter for those where reputation from astrid to ?other is not 30.
 		// Also ensure that the player does not have a spouse
-		var r10 = new DBQuery()
+		query = new DBQuery()
 			.Where( "astrid.relationships.?other" )
 			.Where( "not astrid.relationships.?other.reputation!30" )
 			// below will fail because player has spouse
-			.Where( "not player.relationships.?x.tags.spouse" )
-			.Run( db );
+			.Where( "not player.relationships.?x.tags.spouse" );
 
-		Assert.That( r10.Success, Is.EqualTo( false ) );
+		result = query.Run( _db );
+
+		Assert.That( result.Success, Is.EqualTo( false ) );
 
 		// For all relationships astrid has with an ?other
 		// filter for those where reputation from astrid to ?other is not 30.
 		// Also ensure that the player does not have any friends
-		var r11 = new DBQuery()
+		query = new DBQuery()
 			.Where( "astrid.relationships.?other" )
 			.Where( "not astrid.relationships.?other.reputation!30" )
-			.Where( "not player.relationships.?x.tags.friend" )
+			.Where( "not player.relationships.?x.tags.friend" );
+
+		result = query.Run( _db );
+
+		Assert.That( result.Success, Is.EqualTo( true ) );
+	}
+
+	[Test]
+	public void TestCompoundQueryWithVars()
+	{
+		// Compound query with multiple variables
+		var query = new DBQuery()
+			.Where( "?speaker.relationships.?other.reputation!?r0" )
+			.Where( "gt ?r0 10" )
+			.Where( "player.relationships.?other.reputation!?r1" )
+			.Where( "lt ?r1 0" )
+			.Where( "neq ?speaker player" );
+
+		var result = query.Run( _db );
+
+		Assert.That( result.Success, Is.EqualTo( true ) );
+		Assert.That( result.Bindings.Length, Is.EqualTo( 1 ) );
+	}
+
+	[Test]
+	public void TestResultsDataTypes()
+	{
+		// Compound query with multiple variables
+		var query = new DBQuery()
+			.Where( "?speaker.relationships.?other.reputation!?r0" )
+			.Where( "gt ?r0 10" )
+			.Where( "player.relationships.?other.reputation!?r1" )
+			.Where( "lt ?r1 0" )
+			.Where( "neq ?speaker player" );
+
+		var result = query.Run( _db );
+
+		Assert.That( result.Success, Is.True );
+		Assert.That( result.Bindings.Length, Is.EqualTo( 1 ) );
+		Assert.That( result.Bindings[0]["?speaker"] is string, Is.True );
+		Assert.That( result.Bindings[0]["?other"] is string, Is.True );
+		Assert.That( result.Bindings[0]["?r0"] is int, Is.True );
+		Assert.That( result.Bindings[0]["?r1"] is int, Is.True );
+	}
+
+	/// <summary>
+	/// Test insertion, deletion, and query of string literals
+	/// </summary>
+	[Test]
+	public void TestStringLiteral()
+	{
+		RePraxisDatabase db = new RePraxisDatabase();
+
+		db.Insert( "toph.fullName!Toph Beifong" );
+		db.Insert( "toph.displayName!Toph: The greatest Earthbender in the world" );
+
+		Assert.That( db.Assert( "toph.fullName!Toph Beifong" ) );
+
+		QueryResult queryResult;
+
+		queryResult = new DBQuery()
+			.Where( "toph.fullName!?fullName" )
 			.Run( db );
 
-		Assert.That( r11.Success, Is.EqualTo( true ) );
+		Assert.That( queryResult.Bindings[0]["?fullName"], Is.EqualTo( "Toph Beifong" ) );
 	}
 }